+++ src/governance-compliance/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { ComplianceControlFrameworkDescriptor } from './compliance-control-framework/compliance-control-framework-descriptor.js';
 import { EnterpriseRiskManagementDescriptor } from './enterprise-risk-management/enterprise-risk-management-descriptor.js';
 import { GovernanceOperatingModelDescriptor } from './operating-model/governance-operating-model-descriptor.js';
 import { GovernanceComplianceOverviewDescriptor } from './overview/governance-compliance-overview-descriptor.js';
@@ -13,6 +14,7 @@ export function addGovernanceCompliance(services) {
   services.registerSingleton('GovernanceOperatingModelDescriptor', () => new GovernanceOperatingModelDescriptor());
   services.registerSingleton('PolicyArchitectureDescriptor', () => new PolicyArchitectureDescriptor());
   services.registerSingleton('EnterpriseRiskManagementDescriptor', () => new EnterpriseRiskManagementDescriptor());
+  services.registerSingleton('ComplianceControlFrameworkDescriptor', () => new ComplianceControlFrameworkDescriptor());
 
   return services;
 }